added get local storage

+++ src/lib/utils/storage/useLocalStorage/index.js
@@ -1,23 +1,23 @@
 import { Storage } from '@capacitor/storage'
 import useAsync from '@useweb/use-async'
 
-export default function useLocalStorage({ key } = {}) {
-  const get = useAsync(async () => {
-    const { value } = await Storage.get({ key })
-    const valueParsed = JSON.parse(value)
+export default function useLocalStorage({ action = 'get', key } = {}) {
+  return useAsync(async (data) => {
+    if (action === 'get') {
+      const { value } = await Storage.get({ key })
+      const valueParsed = JSON.parse(value)
 
-    return valueParsed
+      return valueParsed
+    }
+
+    if (action === 'set') {
+      const valueStringifyed = JSON.stringify(data.value)
+
+      await Storage.set({ key: data.key, value: valueStringifyed })
+    }
+
+    if (action === 'remove') {
+      await Storage.remove({ key })
+    }
   })
-
-  const set = useAsync(async (data) => {
-    const valueStringifyed = JSON.stringify(data.value)
-
-    await Storage.set({ key: data.key, value: valueStringifyed })
-  })
-
-  const remove = useAsync(async () => {
-    await Storage.remove({ key })
-  })
-
-  return { get, set, remove }
 }
